+++ background/battle_field.js
@@ -9,7 +9,7 @@ const battleField = {
     switch (method) {
       case 'autoFinish': // old/depreciated message
         battleField.processStartBattle(data, function () {
-          battleField.processBattleMove(data.state)
+          battleField.processBattleMove(data.state, true)
         })
         break
       case 'startPvP': // old/depreciated message
@@ -17,7 +17,7 @@ const battleField = {
         battleField.processStartBattle(data, function () {
           if (data.isAutoBattle) {
             // Battle finished immediately
-            battleField.processBattleMove(data.state)
+            battleField.processBattleMove(data.state, true)
           }
         })
         break
@@ -64,7 +64,7 @@ const battleField = {
     }
   },
 
-  processBattleMove: function (data) {
+  processBattleMove: function (data, isAutoBattle) {
     if (battleField.lastPlayerAttacked < 0) {
       // Unknown player (Expedition, other non-pvp battle, or error), no need to update battle statistics
       return
@@ -93,7 +93,7 @@ const battleField = {
     if (debug) {
       console.log('Battle ' + (battleWon ? 'won' : 'lost') + ', lost HP: ' + lostHP + ' died: ', unitsDied)
     }
-    battleField.storeBattleResults(battleWon, data.surrenderBit === 1, lostHP, unitsDied)
+    battleField.storeBattleResults(battleWon, data.surrenderBit === 1, lostHP, unitsDied, isAutoBattle === true)
   },
 
   getArmies: function (armiesData) {
@@ -175,7 +175,7 @@ const battleField = {
     })
   },
 
-  storeBattleResults: function (battleWon, surrendered, lostHP, unitsDied) {
+  storeBattleResults: function (battleWon, surrendered, lostHP, unitsDied, isAutoBattle) {
     syncGet({ 'playerArmies': {} }, function (result) {
       let playerArmies = result.playerArmies
       let armyDetails = playerArmies[battleField.lastPlayerAttacked] || {}
@@ -188,6 +188,7 @@ const battleField = {
       details.won = battleWon
       details.lost = !battleWon
       details.surrendered = surrendered
+      details.isAutoBattle = isAutoBattle
       armyDetails.battles.details[armyDetails.battles.details.length - 1] = details
 
       armyDetails.battles[battleWon ? 'wins' : 'loses']++